+++ sdk/preview/design/index.js
@@ -1,7 +1,9 @@
 const di = window.canva.designInteraction;
 export const addNativeElement = di.addNativeElement.bind(di);
+export const addPage = di.addPage.bind(di);
 export const addAudioTrack = di.addAudioTrack.bind(di);
 export const getCurrentPageContext = di.getCurrentPageContext.bind(di);
+export const getDefaultPageDimensions = di.getDefaultPageDimensions.bind(di);
 export const initAppElement = di.initAppElement.bind(di);
 
 export const ui = (() => {